Move view logic to multiselect list, use useCallback hooks
